refactor: use badge component with variantes
