+++ src/components/User/Badges/BadgeCard.tsx
@@ -3,14 +3,13 @@ import React from 'react'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
 
-import { Badge } from '../../../entities/Badges/types'
+import { Badge as GovernanceBadge } from '../../../entities/Badges/types'
 
+import Badge from './Badge'
 import './BadgeCard.css'
 
-const NO_IMAGE = require('../../../images/no-image.png').default
-
 interface Props {
-  badge: Badge
+  badge: GovernanceBadge
   onClick: () => void
 }
 
@@ -19,13 +18,11 @@ export default function BadgeCard({ badge, onClick }: Props) {
 
   return (
     <div className="BadgeCard" key={`${badge.name}-id`} onClick={onClick}>
-      <div className="BadgeCard__Icon">
-        <img src={badge.image} onError={(e) => (e.currentTarget.src = NO_IMAGE)} alt="badge-icon" />
-        <div className="BadgeCard__Info">
-          <div className="BadgeCard__Title">{badge.name}</div>
-          <div className="BadgeCard__MintDate">
-            {t('component.badge_card.mint_date', { at: Time.unix(badge.createdAt).fromNow() })}
-          </div>
+      <Badge badge={badge} className="BadgeCard__Icon" />
+      <div className="BadgeCard__Info">
+        <div className="BadgeCard__Title">{badge.name}</div>
+        <div className="BadgeCard__MintDate">
+          {t('component.badge_card.mint_date', { at: Time.unix(badge.createdAt).fromNow() })}
         </div>
       </div>
     </div>
