+++ src/components/User/Badges/BadgeDetail.tsx
@@ -4,23 +4,20 @@ import Markdown from 'decentraland-gatsby/dist/components/Text/Markdown'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
 
-import { Badge } from '../../../entities/Badges/types'
+import { Badge as GovernanceBadge } from '../../../entities/Badges/types'
 
+import Badge, { BadgeVariant } from './Badge'
 import './BadgeDetail.css'
 
-const NO_IMAGE = require('../../../images/no-image.png').default
-
 interface Props {
-  badge: Badge
+  badge: GovernanceBadge
 }
 
 export default function BadgeDetail({ badge }: Props) {
   const t = useFormatMessage()
   return (
     <div className="BadgeDetail__Container">
-      <div className="BadgeDetail__Icon">
-        <img src={badge.image} onError={(e) => (e.currentTarget.src = NO_IMAGE)} alt="badge-icon" />
-      </div>
+      <Badge badge={badge} variant={BadgeVariant.Primary} className="BadgeDetail__Icon" />
       <div className="BadgeDetail__Info">
         <div className="BadgeDetail__Title">{badge.name}</div>
         <div className="BadgeDetail__MintDate">
